[flex_web_service] Fix return value, check for messages
Messages: The function now checks for and reports
messages from IB in the downloaded report. This is
currently only supported (tested) for CSV reports.

Return value: If an attempt to download the report is
made, the value of download.file will be
returned. Otherwise, an error occured, and the return
value will be 1, invisibly.